pyproject.toml
[tool.poetry]
name = "eppr-lambda"
version = "0.0.1"
description = ""
authors = ["ExtendedPPR <extendedppr@gmail.com>"]
readme = "README.md"
packages = [{include="eppr", from="src"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
pymongo = "^4.7.1"
apiflask = "^1.2.3"
flask-apispec = "^0.11.4"
werkzeug = "2.3.8"
boto3 = "1.34.98"
python-dotenv = "^1.0.1"

[tool.poetry.group.test.dependencies]
pytest = "^8.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
